Moved some authentication around
