+++ src/main/java/com/blackducksoftware/integration/jira/config/controller/AccessConfigController.java
@@ -81,7 +81,7 @@ public class AccessConfigController extends ConfigController {
         try {
             final boolean userAvailable = getAuthorizationChecker().isUserAvailable(request);
             if (!userAvailable) {
-                return Response.status(Response.Status.UNAUTHORIZED).build();
+                return Response.status(Response.Status.CREATED).build();
             }
 
             responseObject = executeAsTransaction(() -> accessConfigActions.updateConfigWithJiraGroups(request, adminConfig.getHubJiraGroups()));
