+++ src/test/java/com/blackducksoftware/integration/jira/config/BlackDuckJiraServletTest.java
@@ -27,7 +27,7 @@ import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
 
-import com.blackducksoftware.integration.jira.config.controller.BlackDuckJiraServlet;
+import com.blackducksoftware.integration.jira.config.controller.servlet.BlackDuckJiraServlet;
 import com.blackducksoftware.integration.jira.mocks.HttpServletRequestMock;
 import com.blackducksoftware.integration.jira.mocks.HttpServletResponseMock;
 import com.blackducksoftware.integration.jira.mocks.LoginUriProviderMock;
